+++ src/lib/flags.ts
@@ -13,7 +13,7 @@ const testnetFeatureFlag = (inventory: GameState["inventory"]) =>
  *
  * Do not delete JEST_TEST.
  */
-type FeatureName = "JEST_TEST" | "FRUIT_QUEST" | "PUMPKIN_PLAZA";
+type FeatureName = "JEST_TEST" | "FRUIT_QUEST" | "PUMPKIN_PLAZA" | "DELIVERIES";
 
 type FeatureFlag = (inventory: GameState["inventory"]) => boolean;
 
@@ -21,6 +21,7 @@ const featureFlags: Record<FeatureName, FeatureFlag> = {
   JEST_TEST: defaultFeatureFlag,
   FRUIT_QUEST: testnetFeatureFlag,
   PUMPKIN_PLAZA: defaultFeatureFlag,
+  DELIVERIES: defaultFeatureFlag,
 };
 
 export const hasFeatureAccess = (
